# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Microservice to manage PagoPA Backoffice",
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "SelfCare Backoffice",
    "version" : "1.4.5"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  }, {
    "url" : "https://{host}{basePath}",
    "variables" : {
      "basePath" : {
        "default" : "/backoffice/v1"
      },
      "host" : {
        "default" : "api.dev.platform.pagopa.it",
        "enum" : [ "api.dev.platform.pagopa.it", "api.uat.platform.pagopa.it", "api.platform.pagopa.it" ]
      }
    }
  } ],
  "paths" : {
...
    "/creditor-institutions" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getCreditorInstitutions",
        "parameters" : [ {
          "description" : "Number of elements on one page. Default = 50",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "description" : "Page number. Page value starts from 0",
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Creditor institution code",
          "in" : "query",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Creditor institution name",
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Sorting method for paginated result",
          "in" : "query",
          "name" : "sorting",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get paginated list of creditor institutions",
        "tags" : [ "Creditor institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "createCreditorInstitution",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreditorInstitutionDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionDetailsResource"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Create a new creditor institution",
        "tags" : [ "Creditor institutions" ]
      }
    },
    "/creditor-institutions/{ci-code}" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getCreditorInstitutionDetails",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get the detail of specific creditor institution",
        "tags" : [ "Creditor institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "put" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "updateCreditorInstitutionDetails",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCreditorInstitutionDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Update an existing creditor institution",
        "tags" : [ "Creditor institutions" ]
      }
    },
    "/creditor-institutions/{ci-code}/full" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getBrokerAndEcDetails",
        "parameters" : [ {
          "description" : "Broker code of an EC",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BrokerAndEcDetailsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get PSP broker details",
        "tags" : [ "Creditor institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "createCreditorInstitutionAndBroker",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreditorInstitutionAndBrokerDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionDetailsResource"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Create a new creditor institution that can intermediate as a broker",
        "tags" : [ "Creditor institutions" ]
      }
    },
...
    "/creditor-institutions/{ci-code}/segregation-codes" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getCreditorInstitutionSegregationcodes",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionAssociatedCodeList"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get segregation code associations with creditor institution",
        "tags" : [ "Creditor institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/creditor-institutions/{ci-code}/station" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "associateStationToCreditorInstitution",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreditorInstitutionStationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionStationEditResource"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            },
            "description" : "Bad Request",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            },
            "description" : "Forbidden",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemJson"
                }
              }
            },
            "description" : "Service unavailable",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Creates the relationship between the created station and the creditorInstitution",
        "tags" : [ "Creditor institutions" ]
      }
    },
    "/creditor-institutions/{ci-code}/station/{station-code}" : {
      "delete" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "deleteCreditorInstitutionStationRelationship",
        "parameters" : [ {
          "description" : "Creditor institution code",
          "in" : "path",
          "name" : "ci-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Channlecode or StationCode",
          "in" : "path",
          "name" : "station-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "delete the relationship between the created station and the creditorInstitution",
        "tags" : [ "Creditor institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/institutions" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getInstitutions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InstitutionDetail"
                  }
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Retrieves all the onboarded institutions related to the logged user",
        "tags" : [ "Institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/institutions/{institution-id}" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getInstitution",
        "parameters" : [ {
          "description" : "Institution's unique internal identifier",
          "in" : "path",
          "name" : "institution-id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Institution"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Retrieves the details of an institution",
        "tags" : [ "Institutions" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
    "/stations/{station-code}/creditor-institutions" : {
      "get" : {
        "description" : "Internal | External | Synchronous | Authorization | Authentication | TPS | Idempotency | Stateless | Read/Write Intense | Cacheable\n-|-|-|-|-|-|-|-|-|-\nY | N | Y | JWT | JWT | 1.0/sec | Y | Y |  | N\n",
        "operationId" : "getCreditorInstitutionsByStationCode",
        "parameters" : [ {
          "description" : "Station Code",
          "in" : "path",
          "name" : "station-code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Filter by creditor institution name",
          "in" : "query",
          "name" : "ci-name",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditorInstitutionsResource"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "JWT" : [ ]
        } ],
        "summary" : "Get Creditor Institutions By Station Code",
        "tags" : [ "Stations" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreditorInstitutionDetailsResource" : {
        "required" : [ "businessName", "creditorInstitutionCode", "enabled" ],
        "type" : "object",
        "properties" : {
          "address" : {
            "$ref" : "#/components/schemas/CreditorInstitutionAddressResource"
          },
          "applicationCode" : {
            "type" : "string"
          },
          "auxDigit" : {
            "type" : "string"
          },
          "broadcast" : {
            "type" : "boolean"
          },
          "businessName" : {
            "type" : "string"
          },
          "creditorInstitutionCode" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "mod4" : {
            "type" : "string"
          },
          "pspPayment" : {
            "type" : "boolean"
          },
          "reportingFtp" : {
            "type" : "boolean"
          },
          "reportingZip" : {
            "type" : "boolean"
          },
          "segregationCode" : {
            "type" : "string"
          }
        }
      },
      "CreditorInstitutionDto" : {
        "required" : [ "address", "businessName", "creditorInstitutionCode", "enabled", "pspPayment", "reportingFtp", "reportingZip" ],
        "type" : "object",
        "properties" : {
          "address" : {
            "$ref" : "#/components/schemas/CreditorInstitutionAddressDto"
          },
          "businessName" : {
            "type" : "string"
          },
          "creditorInstitutionCode" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "pspPayment" : {
            "type" : "boolean"
          },
          "reportingFtp" : {
            "type" : "boolean"
          },
          "reportingZip" : {
            "type" : "boolean"
          }
        }
      },
...
      "CreditorInstitutionsResource" : {
        "type" : "object",
        "properties" : {
          "creditor_institutions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreditorInstitutionResource"
            }
          },
          "page_info" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "JWT" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token get after Login",
        "scheme" : "bearer",
        "type" : "http"
      },
      "SubKey" : {
        "description" : "The Azure Subscription Key to access this API.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}